refactor: Extract session provider into separate file
