+++ src/app/provider.tsx
@@ -1,0 +1,14 @@
+"use client";
+import { SessionProvider } from "next-auth/react";
+
+export const CustomProviders = ({
+  children,
+}: {
+  children: React.ReactNode
+}) => {
+  return (
+    <SessionProvider>
+        {children}
+    </SessionProvider>
+  )
+}
