--- src/lib/react-query/QueryClientProvider.tsx
@@ -1,14 +1,0 @@
-"use client";
-import { SessionProvider } from "next-auth/react";
-
-export const CustomProviders = ({
-  children,
-}: {
-  children: React.ReactNode
-}) => {
-  return (
-    <SessionProvider>
-        {children}
-    </SessionProvider>
-  )
-}
