--- src/lib/react-query/index.ts
@@ -1,1 +1,0 @@
-export * from './QueryClientProvider';
